refactor: マップ関連の責務をPluginに移行
