+++ Bakajinrou-Plugin/src/main/kotlin/com/github/tanokun/bakajinrou/plugin/map/GameMap.kt
@@ -1,6 +1,7 @@
-package com.github.tanokun.bakajinrou.api.map
+package com.github.tanokun.bakajinrou.plugin.map
 
 import kotlinx.serialization.Serializable
+import org.bukkit.Material
 import kotlin.time.Duration
 
 @Serializable
@@ -9,4 +10,5 @@ data class GameMap(
     val spawnPoint: PointLocation,
     val lobbyPoint: PointLocation,
     val startTime: Duration,
+    val icon: Material
 )
