+++ Bakajinrou-Plugin/src/main/kotlin/com/github/tanokun/bakajinrou/plugin/map/result/MapDeletionResult.kt
@@ -1,6 +1,6 @@
-package com.github.tanokun.bakajinrou.game.map.result
+package com.github.tanokun.bakajinrou.plugin.map.result
 
-import com.github.tanokun.bakajinrou.api.map.GameMap
+import com.github.tanokun.bakajinrou.plugin.map.GameMap
 
 /**
  * マップを削除した時の、結果を表すオブジェクトです。
